Safari-609.1.11.1/LayoutTests/http/tests/referrer-policy-iframe/origin-when-cross-origin/same-origin-expected.txt

Tests the behavior of origin-when-cross-origin referrer policy when same origin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is "http://127.0.0.1:8000/referrer-policy-iframe/origin-when-cross-origin/same-origin.html"
PASS successfullyParsed is true

TEST COMPLETE
